Add refactoring stage I
