+++ pkg/controller/utils/common/common.go
@@ -1,6 +1,13 @@
 package common
 
-import "os"
+import (
+	"github.com/go-logr/logr"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
+	"net"
+	"os"
+	"strings"
+)
 
 // GetEnvWithDefault return GetEnv(name) if exists else return defVal
 func GetEnvWithDefault(name, defVal string) string {
@@ -9,6 +16,15 @@ func GetEnvWithDefault(name, defVal string) string {
 		return str
 	}
 	return defVal
+}
+
+func GetEnvVarIndex(envVarName string, env *[]corev1.EnvVar) int {
+	for i, value := range *env {
+		if value.Name == envVarName {
+			return i
+		}
+	}
+	return 0
 }
 
 func Contains(list []string, s string) bool {
@@ -27,4 +43,22 @@ func Remove(list []string, s string) []string {
 		}
 	}
 	return list
+}
+
+func ToMilliDecimalQuantity(value int64) resource.Quantity {
+	return *resource.NewMilliQuantity(value, resource.DecimalSI)
+}
+
+func GetURISchemeProtocol(URIScheme corev1.URIScheme) string {
+	return strings.ToLower(string(URIScheme))
+}
+
+func LookupHost(host string, logger logr.Logger) (string, error) {
+	addresses, err := net.LookupHost(host)
+	if err != nil {
+		logger.Error(err, "host does not resolve")
+		return "", err
+	}
+	logger.Info("host resolved", "host", host, "addresses", addresses)
+	return host, nil
 }
