+++ pkg/controller/utils/common/labels.go
@@ -1,0 +1,10 @@
+package common
+
+// LabelsResource returns the labels that must me applied to the resource
+func LabelsResource(name, resourceType string) map[string]string {
+	m := map[string]string{"infinispan_cr": name, "clusterName": name}
+	if resourceType != "" {
+		m["app"] = resourceType
+	}
+	return m
+}
