+++ pkg/controller/utils/infinispan/configuration/configmap.go
@@ -1,0 +1,91 @@
+package configuration
+
+import (
+	"context"
+	"github.com/go-logr/logr"
+	"k8s.io/apimachinery/pkg/runtime"
+	"strings"
+
+	infinispanv1 "github.com/infinispan/infinispan-operator/pkg/apis/infinispan/v1"
+	"github.com/infinispan/infinispan-operator/pkg/controller/utils/common"
+	"gopkg.in/yaml.v2"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+)
+
+func ConfigMapForInfinispan(xsite *XSite, m *infinispanv1.Infinispan, client client.Client, scheme *runtime.Scheme, logger logr.Logger) (*corev1.ConfigMap, error) {
+	name := m.ObjectMeta.Name
+	namespace := m.ObjectMeta.Namespace
+
+	loggingCategories := m.CopyLoggingCategories()
+	config := xsite.CreateInfinispanConfiguration(name, loggingCategories, namespace)
+	err := setupConfigForEncryption(m, &config, client, logger)
+	if err != nil {
+		return nil, err
+	}
+	configYaml, err := yaml.Marshal(config)
+	if err != nil {
+		return nil, err
+	}
+	lsConfigMap := common.LabelsResource(m.Name, "infinispan-configmap-configuration")
+	configMap := &corev1.ConfigMap{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "v1",
+			Kind:       "ConfigMap",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name + "-configuration",
+			Namespace: namespace,
+			Labels:    lsConfigMap,
+		},
+		Data: map[string]string{"infinispan.yaml": string(configYaml)},
+	}
+
+	// Set Infinispan instance as the owner and controller
+	controllerutil.SetControllerReference(m, configMap, scheme)
+	return configMap, nil
+}
+
+func setupConfigForEncryption(m *infinispanv1.Infinispan, c *InfinispanConfiguration, client client.Client, logger logr.Logger) error {
+	ee := m.Spec.Security.EndpointEncryption
+	if ee.Type == "service" {
+		if strings.Contains(ee.CertServiceName, "openshift.io") {
+			c.Keystore.CrtPath = "/etc/encrypt"
+			c.Keystore.Path = "/opt/infinispan/server/conf/keystore"
+			c.Keystore.Password = "password"
+			c.Keystore.Alias = "server"
+			return nil
+		}
+	}
+	// Fetch the tls secret if name is provided
+	tlsSecretName := m.GetEncryptionSecretName()
+	if tlsSecretName != "" {
+		tlsSecret := &corev1.Secret{}
+		err := client.Get(context.TODO(), types.NamespacedName{Name: tlsSecretName, Namespace: m.Namespace}, tlsSecret)
+		if err != nil {
+			if errors.IsNotFound(err) {
+				logger.Error(err, "Secret %s for endpoint encryption not found.", tlsSecretName)
+				return err
+			}
+			logger.Error(err, "Error in getting secret %s for endpoint encryption.", tlsSecretName)
+			return err
+		}
+		if _, ok := tlsSecret.Data["keystore.p12"]; ok {
+			// If user provide a keystore in secret then use it ...
+			c.Keystore.Path = "/etc/encrypt/keystore.p12"
+			c.Keystore.Password = string(tlsSecret.Data["password"])
+			c.Keystore.Alias = string(tlsSecret.Data["alias"])
+		} else {
+			// ... else suppose tls.key and tls.crt are provided
+			c.Keystore.CrtPath = "/etc/encrypt"
+			c.Keystore.Path = "/opt/infinispan/server/conf/keystore"
+			c.Keystore.Password = "password"
+			c.Keystore.Alias = "server"
+		}
+	}
+	return nil
+}
