+++ pkg/controller/utils/infinispan/configuration/deployment.go
@@ -1,0 +1,214 @@
+package configuration
+
+import (
+	"encoding/json"
+	"os"
+	"time"
+
+	"github.com/go-logr/logr"
+	infinispanv1 "github.com/infinispan/infinispan-operator/pkg/apis/infinispan/v1"
+	"github.com/infinispan/infinispan-operator/pkg/controller/constants"
+	comutil "github.com/infinispan/infinispan-operator/pkg/controller/utils/common"
+	ispncluster "github.com/infinispan/infinispan-operator/pkg/controller/utils/infinispan"
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+)
+
+// DeploymentForInfinispan returns an infinispan Deployment object
+func DeploymentForInfinispan(m *infinispanv1.Infinispan, secret *corev1.Secret, config *corev1.ConfigMap, scheme *runtime.Scheme, logger logr.Logger) (*appsv1.StatefulSet, error) {
+	reqLogger := logger.WithValues("Request.Namespace", m.Namespace, "Request.Name", m.Name)
+	lsPod := comutil.LabelsResource(m.Name, "infinispan-pod")
+	// This field specifies the flavor of the
+	// Infinispan cluster. "" is plain community edition (vanilla)
+
+	memory := resource.MustParse(m.Spec.Container.Memory)
+	cpuRequests, cpuLimits := m.GetCpuResources()
+	javaOptions, err := m.GetJavaOptions()
+	if err != nil {
+		return nil, err
+	}
+
+	envVars := []corev1.EnvVar{
+		{Name: "CONFIG_PATH", Value: "/etc/config/infinispan.yaml"},
+		{Name: "IDENTITIES_PATH", Value: "/etc/security/identities.yaml"},
+		{Name: "JAVA_OPTIONS", Value: javaOptions},
+		{Name: "EXTRA_JAVA_OPTIONS", Value: m.Spec.Container.ExtraJvmOpts},
+		{Name: "DEFAULT_IMAGE", Value: constants.DefaultImageName},
+	}
+
+	// Adding additional variables listed in ADDITIONAL_VARS env var
+	envVar, defined := os.LookupEnv("ADDITIONAL_VARS")
+	if defined {
+		var addVars []string
+		err := json.Unmarshal([]byte(envVar), &addVars)
+		if err == nil {
+			for _, name := range addVars {
+				value, defined := os.LookupEnv(name)
+				if defined {
+					envVars = append(envVars, corev1.EnvVar{Name: name, Value: value})
+				}
+			}
+		}
+	}
+	replicas := m.Spec.Replicas
+	protocolScheme := m.GetEndpointScheme()
+	dep := &appsv1.StatefulSet{
+
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "apps/v1",
+			Kind:       "StatefulSet",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:        m.ObjectMeta.Name,
+			Namespace:   m.ObjectMeta.Namespace,
+			Annotations: constants.DeploymentAnnotations,
+			Labels:      map[string]string{"template": "infinispan-ephemeral"},
+		},
+		Spec: appsv1.StatefulSetSpec{
+			UpdateStrategy: appsv1.StatefulSetUpdateStrategy{Type: appsv1.RollingUpdateStatefulSetStrategyType},
+			Selector: &metav1.LabelSelector{
+				MatchLabels: lsPod,
+			},
+			Replicas: &replicas,
+			Template: corev1.PodTemplateSpec{
+				ObjectMeta: metav1.ObjectMeta{
+					Labels:      lsPod,
+					Annotations: map[string]string{"updateDate": time.Now().String()},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{{
+						Image: m.Spec.Image,
+						Name:  "infinispan",
+						Env:   envVars,
+						LivenessProbe: &corev1.Probe{
+							Handler:             ispncluster.ClusterStatusHandler(protocolScheme),
+							FailureThreshold:    5,
+							InitialDelaySeconds: 10,
+							PeriodSeconds:       60,
+							SuccessThreshold:    1,
+							TimeoutSeconds:      80},
+						Ports: []corev1.ContainerPort{
+							{ContainerPort: 8888, Name: "ping", Protocol: corev1.ProtocolTCP},
+							{ContainerPort: 11222, Name: "hotrod", Protocol: corev1.ProtocolTCP},
+						},
+						ReadinessProbe: &corev1.Probe{
+							Handler:             ispncluster.ClusterStatusHandler(protocolScheme),
+							FailureThreshold:    5,
+							InitialDelaySeconds: 10,
+							PeriodSeconds:       10,
+							SuccessThreshold:    1,
+							TimeoutSeconds:      80},
+						Resources: corev1.ResourceRequirements{
+							Requests: corev1.ResourceList{
+								"cpu":    cpuRequests,
+								"memory": memory,
+							},
+							Limits: corev1.ResourceList{
+								"cpu":    cpuLimits,
+								"memory": memory,
+							},
+						},
+						VolumeMounts: []corev1.VolumeMount{{
+							Name:      "config-volume",
+							MountPath: "/etc/config",
+						}, {
+							Name:      "identities-volume",
+							MountPath: "/etc/security",
+						}},
+					}},
+					Volumes: []corev1.Volume{{
+						Name: "config-volume",
+						VolumeSource: corev1.VolumeSource{
+							ConfigMap: &corev1.ConfigMapVolumeSource{
+								LocalObjectReference: corev1.LocalObjectReference{Name: config.Name},
+							},
+						},
+					}, {
+						Name: "identities-volume",
+						VolumeSource: corev1.VolumeSource{
+							Secret: &corev1.SecretVolumeSource{
+								SecretName: secret.Name,
+							},
+						},
+					}},
+				},
+			},
+		},
+	}
+
+	// Persistent vol size must exceed memory size
+	// so that it can contain all the in memory data
+	pvSize := constants.DefaultPVSize
+	if pvSize.Cmp(memory) < 0 {
+		pvSize = memory
+	}
+
+	if m.IsDataGrid() && m.Spec.Service.Container.Storage != "" {
+		var pvErr error
+		pvSize, pvErr = resource.ParseQuantity(m.Spec.Service.Container.Storage)
+		if pvErr != nil {
+			return nil, pvErr
+		}
+		if pvSize.Cmp(memory) < 0 {
+			reqLogger.Info("WARNING: persistent volume size is less than memory size. Graceful shutdown may not work.", "Volume Size", pvSize, "Memory", memory)
+		}
+	}
+
+	dep.Spec.VolumeClaimTemplates = []corev1.PersistentVolumeClaim{{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      m.ObjectMeta.Name,
+			Namespace: m.ObjectMeta.Namespace,
+		},
+		Spec: corev1.PersistentVolumeClaimSpec{
+			AccessModes: []corev1.PersistentVolumeAccessMode{
+				"ReadWriteOnce",
+			},
+			Resources: corev1.ResourceRequirements{
+				Requests: corev1.ResourceList{
+					corev1.ResourceStorage: pvSize,
+				},
+			},
+		}}}
+
+	// Adding persistent volume mount
+	v := &dep.Spec.Template.Spec.Containers[0].VolumeMounts
+	*v = append(*v, corev1.VolumeMount{
+		Name:      m.ObjectMeta.Name,
+		MountPath: "/opt/infinispan/server/data",
+	})
+	// Adding an init container that run chmod if needed
+	if chmod, ok := os.LookupEnv("MAKE_DATADIR_WRITABLE"); ok && chmod == "true" {
+		dep.Spec.Template.Spec.InitContainers = []corev1.Container{{
+			Image:   "busybox",
+			Name:    "chmod-pv",
+			Command: []string{"sh", "-c", "chmod -R g+w /opt/infinispan/server/data"},
+			VolumeMounts: []corev1.VolumeMount{{
+				Name:      m.ObjectMeta.Name,
+				MountPath: "/opt/infinispan/server/data",
+			}},
+		}}
+	}
+	setupVolumesForEncryption(m, dep)
+
+	// Set Infinispan instance as the owner and controller
+	controllerutil.SetControllerReference(m, dep, scheme)
+	return dep, nil
+}
+
+func setupVolumesForEncryption(m *infinispanv1.Infinispan, dep *appsv1.StatefulSet) {
+	secretName := m.GetEncryptionSecretName()
+	if secretName != "" {
+		v := &dep.Spec.Template.Spec.Volumes
+		vm := &dep.Spec.Template.Spec.Containers[0].VolumeMounts
+		*vm = append(*vm, corev1.VolumeMount{Name: "encrypt-volume", MountPath: "/etc/encrypt"})
+		*v = append(*v, corev1.Volume{Name: "encrypt-volume",
+			VolumeSource: corev1.VolumeSource{
+				Secret: &corev1.SecretVolumeSource{
+					SecretName: secretName,
+				}}})
+	}
+}
