+++ pkg/controller/utils/infinispan/configuration/secret_identities.go
@@ -1,0 +1,38 @@
+package configuration
+
+import (
+	infinispanv1 "github.com/infinispan/infinispan-operator/pkg/apis/infinispan/v1"
+	comutil "github.com/infinispan/infinispan-operator/pkg/controller/utils/common"
+	"github.com/infinispan/infinispan-operator/pkg/controller/utils/k8s"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+)
+
+func SecretForInfinispan(identities []byte, m *infinispanv1.Infinispan, scheme *runtime.Scheme) *corev1.Secret {
+	lsSecret := comutil.LabelsResource(m.Name, "infinispan-secret-identities")
+	secretName := m.GetSecretName()
+	secret := &corev1.Secret{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "v1",
+			Kind:       "Secret",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      secretName,
+			Namespace: m.ObjectMeta.Namespace,
+			Labels:    lsSecret,
+		},
+		Type:       corev1.SecretType("Opaque"),
+		StringData: map[string]string{"identities.yaml": string(identities)},
+	}
+
+	// Set Infinispan instance as the owner and controller
+	controllerutil.SetControllerReference(m, secret, scheme)
+	return secret
+}
+
+func FindSecret(m *infinispanv1.Infinispan, kubernetes *k8s.Kubernetes) (*corev1.Secret, error) {
+	secretName := m.GetSecretName()
+	return kubernetes.GetSecret(secretName, m.Namespace)
+}
