+++ pkg/controller/utils/infinispan/configuration/service_core.go
@@ -1,0 +1,39 @@
+package configuration
+
+import (
+	infinispanv1 "github.com/infinispan/infinispan-operator/pkg/apis/infinispan/v1"
+	comutil "github.com/infinispan/infinispan-operator/pkg/controller/utils/common"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+)
+
+func ServiceForInfinispan(m *infinispanv1.Infinispan, scheme *runtime.Scheme) *corev1.Service {
+	lsPodSelector := comutil.LabelsResource(m.Name, "infinispan-pod")
+	lsService := comutil.LabelsResource(m.Name, "infinispan-service")
+	service := &corev1.Service{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "v1",
+			Kind:       "Service",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      m.Name,
+			Namespace: m.Namespace,
+			Labels:    lsService,
+		},
+		Spec: corev1.ServiceSpec{
+			Type:     corev1.ServiceTypeClusterIP,
+			Selector: lsPodSelector,
+			Ports: []corev1.ServicePort{
+				{
+					Port: 11222,
+				},
+			},
+		},
+	}
+
+	// Set Infinispan instance as the owner and controller
+	controllerutil.SetControllerReference(m, service, scheme)
+	return service
+}
